+++ module-assembler-intellij-plugin/src/main/java/com/electrolytej/assembler/compose/FieldText.kt
@@ -1,4 +1,4 @@
-package com.electrolytej.manager
+package com.electrolytej.assembler.compose
 
 import com.intellij.ui.components.JBTextField